pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "card-collector"
version = "0.1.0"
description = "Discord bot for collectible card management"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Card Collector", email = "cards@example.com"},
]
dependencies = [
    "discord.py>=2.3.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "pillow>=10.0.0",
    "python-jose[cryptography]>=3.3.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[tool.black]
target-version = ["py311"]
line-length = 88

[tool.ruff]
target-version = "py311"
line-length = 88